Seperate functions and routes
