+++ server.js
@@ -2,6 +2,14 @@ import express from "express";
 import cors from "cors";
 import expressListEndpoints from "express-list-endpoints";
 import { topics } from "./data/topics.json";
+import {
+  extractParam,
+  extractTopicNames,
+  findTopicByName,
+  findSubTopics,
+  findTopicsByKeyword,
+  findQuestionsByLevel,
+} from "./functions/topicFunctions.js";
 
 // Defines the port the app will run on. Defaults to 8080, but can be overridden
 // when starting the server. Example command to overwrite PORT env variable value:
@@ -14,57 +22,7 @@ let app = express();
 app.use(cors());
 app.use(express.json());
 
-// ____________________ FUNCTIONS ____________________
-
-// Function to extract param
-const extractParam = (req, param) => {
-  return req.params[param];
-};
-
-// Function to extract all topic names
-const extractTopicNames = () => topics.map((topic) => topic.name);
-
-// Function to find a topic name
-const findTopicByName = (topicName) => {
-  return topics.find((topic) => topic.name === topicName);
-};
-
-// Function to find topic with specified name
-const findSubTopics = (topicName) => {
-  const topic = topics.find((topic) => topic.name === topicName);
-
-  // If topic is found then return it´s subtopic, if not return null
-  return topic ? topic.subtopics : null;
-};
-
-const findTopicsByKeyword = (querySearch) => {
-  return topics.filter((topic) =>
-    topic.name.toLowerCase().includes(querySearch.toLowerCase())
-  );
-};
-
-const findQuestionsByLevel = (level) => {
-  // Create an array to store questions
-  let questions = [];
-  const limit = 4;
-
-  // Loop through each topic
-  topics.forEach((topic) => {
-    // Loop through each subtopic
-    topic.subtopics.forEach((subtopic) => {
-      // Filter questions based on the level
-      const filteredQuestions = subtopic.questions.filter(
-        (question) => question.level === level
-      );
-      // Add filtered questions to the array
-      questions = questions.concat(filteredQuestions);
-    });
-  });
-
-  return questions.slice(0, limit);
-};
-
-// ____________________ ROUTES ____________________
+// ================== ROUTES ==================
 
 // // Start defining your routes here
 app.route("/").get((req, res) => {
@@ -77,7 +35,7 @@ app.get("/topics", (req, res) => {
   // topics?topic=javascript
   const querySearch = req.query.topic;
   if (querySearch) {
-    const filterTopics = findTopicsByKeyword(querySearch);
+    const filterTopics = findTopicsByKeyword(topics, querySearch);
 
     if (filterTopics.length > 0) {
       res.json(filterTopics);
@@ -92,7 +50,7 @@ app.get("/topics", (req, res) => {
 
 // Get all topic names
 app.get("/topics/name", (req, res) => {
-  const topicNames = extractTopicNames();
+  const topicNames = extractTopicNames(topics);
   res.json(topicNames);
 });
 
@@ -102,7 +60,7 @@ app.get("/topics/:name", (req, res) => {
   const topicName = extractParam(req, "name");
 
   // Passing topics and topics name as params
-  const topic = findTopicByName(topicName);
+  const topic = findTopicByName(topics, topicName);
 
   if (topic) {
     res.json(topic);
@@ -116,7 +74,7 @@ app.get("/topics/:name", (req, res) => {
 app.get("/topics/:name/subtopics", (req, res) => {
   const topicName = extractParam(req, "name");
   // Passing topics and topics name as params
-  const subtopics = findSubTopics(topicName);
+  const subtopics = findSubTopics(topics, topicName);
 
   if (subtopics) {
     res.json(subtopics);
@@ -129,7 +87,7 @@ app.get("/topics/:name/subtopics", (req, res) => {
 app.get("/topics/:name/questions", (req, res) => {
   const topicName = extractParam(req, "name");
   // Passing topic name as params to find topics by name
-  const topic = findTopicByName(topicName);
+  const topic = findTopicByName(topics, topicName);
 
   if (topic) {
     res.json(topic);
@@ -144,7 +102,7 @@ app.get("/topics/:name/questions", (req, res) => {
 app.get("/topics/subtopics/:level/questions", (req, res) => {
   const level = extractParam(req, "level");
 
-  const questions = findQuestionsByLevel(level);
+  const questions = findQuestionsByLevel(topics, level);
   res.json(questions);
 });
